requirements.txt
pandas
numpy
scikit-learn
requests
load_dotenv